pyproject.toml
[tool.poetry]
name = "celerycoursework"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
flake8 = "^7.1.1"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
celery = "^5.4.0"
redis = "^5.0.8"
requests = "^2.32.3"
drf-yasg = "^1.21.7"
django-celery-beat = "^2.7.0"
setuptools = "^74.1.2"
django-cors-headers = "^4.4.0"
black = "^24.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
